Ally Boost Refactor (#14)
* Added support for cheats. These must be enabled using a compiler define symbol: ENABLE_CHEATS. When enabled, saving of individual and team scores is disabled, EXP gain is disabled, and players start with 5 Ally Boost tokens.

* Added CHEAT display to version number in Main Menu, if cheats are enabled.
Enabling cheats now causes Turbo meter to start full during gameplay.

* WIP major refactor of Ally Boost system to support Network games.
Fixed two secrets using the same input code

* Fixed Ally Boosts not being applied when playing a local multiplayer game after leaving a Network game.
Changed starting Ally Boost tokens to 10 when cheats are enabled.
Added "Goals and Rivals" wiki page, including images
Added missing image to "Ally Boost" wiki page.

* Disable ENABLE_CHEATS flag by default

+++ Assets/Scripts/Input/HitResult.cs
@@ -30,7 +30,7 @@ public class HitResult : INetworkSerializable
     /// <summary>
     /// In Network games, this gets or sets the Net ID of the player who hit (or missed) this note.
     /// </summary>
-    public int NetId;
+    public ulong NetId;
 
     /// <summary>
     /// Gets or sets which player hit (or missed) this note.
